pyproject.toml
[project]
name = "salesforce-archivist"
version = "0.1.0"
description = "Salesforce files archiving tool"
authors = [
  {name = "piotrekkr", email = "piotrekkr@users.noreply.github.com"},
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "humanize>=4.12.3",
    "pydantic>=2.11.7",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "simple-salesforce>=1.12.9",
    "types-pyyaml>=6.0.12.20250516",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
archivist = 'salesforce_archivist.cli:cli'

[tool.uv]
dev-dependencies = [
    "mypy>=1.17.0",
    "poethepoet>=0.36.0",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.5",
    "types-requests>=2.32.4.20250611",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "test",
]

[tool.mypy]
files = "src"
disallow_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
no_implicit_optional = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.poe.tasks.test]
help = "Run tests"
cmd  = "pytest ."

[tool.poe.tasks.lint-check]
help = "Check if code is linted"
cmd  = "ruff check"

[tool.poe.tasks.lint]
help = "Lint code"
cmd  = "ruff check --fix"

[tool.poe.tasks.format]
help = "Format code"
cmd  = "ruff format"

[tool.poe.tasks.format-check]
help = "Check code formating"
cmd  = "ruff format --check"

[tool.poe.tasks.mypy]
help = "Run static analysis"
cmd = "mypy"

[tool.poe.tasks.qt]
help = "Run code quality tools"
sequence  = ["format", "lint", "mypy"]
